Fix: 파일 구조 간소화
