--- CodeStarterCamp_Week5/TalentedPersonWithBadPersonality.swift
@@ -1,23 +1,0 @@
-//
-//  TalentedPersonWithBadPersonality.swift
-//  CodeStarterCamp_Week5
-//
-//  Created by 비모 on 2023/03/21.
-//
-
-import Foundation
-
-class TalentedPersonWithBadPersonality: Person, Talent, BadPersonality {
-    var singing: Level?
-    var dancing: Level?
-    var acting: Level?
-    var frequancyOfCursing: Level?
-    
-    init(name: String, height: Int, singing: Level?, dancing: Level?, acting: Level?, frequancyOfCursing: Level?) {
-        super.init(name: name, height: height)
-        self.singing = singing
-        self.dancing = dancing
-        self.acting = acting
-        self.frequancyOfCursing = frequancyOfCursing
-    }
-}
